+++ examples/pod-terminator-operator/client/k8s/clientset/versioned/typed/chaos/v1alpha1/fake/fake_podterminator.go
@@ -8,7 +8,6 @@ import (
 	v1alpha1 "github.com/spotahome/kooper/examples/pod-terminator-operator/v2/apis/chaos/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
-	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	types "k8s.io/apimachinery/pkg/types"
 	watch "k8s.io/apimachinery/pkg/watch"
 	testing "k8s.io/client-go/testing"
@@ -19,9 +18,9 @@ type FakePodTerminators struct {
 	Fake *FakeChaosV1alpha1
 }
 
-var podterminatorsResource = schema.GroupVersionResource{Group: "chaos.spotahome.com", Version: "v1alpha1", Resource: "podterminators"}
+var podterminatorsResource = v1alpha1.SchemeGroupVersion.WithResource("podterminators")
 
-var podterminatorsKind = schema.GroupVersionKind{Group: "chaos.spotahome.com", Version: "v1alpha1", Kind: "PodTerminator"}
+var podterminatorsKind = v1alpha1.SchemeGroupVersion.WithKind("PodTerminator")
 
 // Get takes name of the podTerminator, and returns the corresponding podTerminator object, and an error if there is any.
 func (c *FakePodTerminators) Get(ctx context.Context, name string, options v1.GetOptions) (result *v1alpha1.PodTerminator, err error) {