+++ hack/scripts/integration-test-kind.sh
@@ -3,17 +3,16 @@
 set -o errexit
 set -o nounset
 
-KUBERNETES_VERSION=v${KUBERNETES_VERSION:-1.15.7}
-current_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+KUBERNETES_VERSION=v${KUBERNETES_VERSION:-1.27.3}
+current_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 PREVIOUS_KUBECTL_CONTEXT=$(kubectl config current-context) || PREVIOUS_KUBECTL_CONTEXT=""
 
 function cleanup {
-    if [ ! -z $PREVIOUS_KUBECTL_CONTEXT ]
-    then
-      kubectl config use-context $PREVIOUS_KUBECTL_CONTEXT
-    fi
-    echo "=> Removing kind cluster"
-    kind delete cluster
+  if [ ! -z $PREVIOUS_KUBECTL_CONTEXT ]; then
+    kubectl config use-context $PREVIOUS_KUBECTL_CONTEXT
+  fi
+  echo "=> Removing kind cluster"
+  kind delete cluster
 }
 trap cleanup EXIT